the "all" command
